+++ pkg/ansible/controller/reconcile_test.go
@@ -46,7 +46,6 @@ func TestReconcile(t *testing.T) {
 		Name            string
 		GVK             schema.GroupVersionKind
 		ReconcilePeriod time.Duration
-		ManageStatus    bool
 		Runner          runner.Runner
 		EventHandlers   []events.EventHandler
 		Client          client.Client
@@ -54,6 +53,7 @@ func TestReconcile(t *testing.T) {
 		Result          reconcile.Result
 		Request         reconcile.Request
 		ShouldError     bool
+		ManageStatus    bool
 	}{
 		{
 			Name:            "cr not found",