+++ test/e2e/tls_util_test.go
@@ -22,7 +22,7 @@ import (
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
 	tlsutil "github.com/operator-framework/operator-sdk/pkg/tls"
 
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	apiErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"